pyproject.toml
[project]
name = "etf-grid-design"
version = "0.1.0"
description = "ETF网格交易策略参数设计工具"
authors = [
    {name = "ETF Grid Design", email = "jorbenzhu@gmail.com"}
]
dependencies = [
    "flask>=2.3.0",
    "flask-cors>=4.0.0",
    "tushare>=1.2.89",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "yfinance>=0.2.18",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scipy>=1.11.0",
    "cachetools>=5.3.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "isort>=5.12.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.hatch.build.targets.sdist]
include = [
    "backend/",
    "frontend/",
    "*.md",
    "*.txt",
    ".env.example",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88
